Typedapi as additional competition
Hi all,
I just stumbled upon this project and realized the lack of yet another competitor, [Typedapi](https://github.com/pheymann/typedapi) :). As of now, this is probably the closed you can get to have Haskell's Servant in Scala and it ticks many of the points you have on your list. I would even go so far as to say the no-macros point could be marked as supported, but this is debatable.

Have a look at it and let me know if you think it makes sense to put it into the list.